api/pyproject.toml
[project]
name = "rag-agent-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi==0.135.3",
    "langchain==1.2.15",
    "langchain-chroma==1.1.0",
    "langchain-community==0.4.1",
    "langchain-huggingface==1.2.1",
    "pymupdf==1.27.2.2",
    "python-dotenv==1.2.2",
    "python-multipart==0.0.24",
    "sentence-transformers==5.4.0",
    "torch==2.9.1",
    "uvicorn==0.44.0",
]

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[tool.uv.sources]
torch = { index = "pytorch-cpu" }
